Change default behavior to type during dictation
The default behavior is now to type while dictating
instead of waiting until the end and typing all text in.

The old behavior can be restored using `--defer-output`,
which is implied with `--output=STDOUT`.

There is also a new `--progressive-continuous` option,
say dictation can be left enabled without accumulating the entirety
of words spoken for processing on every update.