Prevent recording until the Vosk model is done loading.
Moved recording_proc_start() below Vosk initialization so it will not record
and buffer audio until after Vosk is ready. This prevents a perceived
sluggishness during the first few moments after a large model is loaded since
there is no rush on the CPU when backlogged audio is passed to Vosk.

When using --verbose, the user will know that you talk because it prints
"Model loaded"

Signed-off-by: Eric Wheeler <vosk-git@z.ewheeler.org>