Add lexicon graph support
A lexicon graph is a feature built into Vosk which allows you to define the
words and phrases that will be recognized be the speech engine. By reducing the
grammar to a specific set of words or commands, recognition becomes almost 100%
accurate.

This patch adds the option --vosk-grammar-file and passes the raw JSON content
to Vosk.  According to Alpha Cephei, you must have a voice model with the
lexicon graph.  I used vosk-model-en-us-0.22-lgraph which works quite well.

More info:
	https://alphacephei.com/nsh/2020/03/27/lookahead.html

See `vosk_recognizer_new_grm` in the API reference:
	https://github.com/alphacep/vosk-api/blob/master/src/vosk_api.h

This is hooked into Python through the Vosk API, here:
	https://github.com/alphacep/vosk-api/blob/fff40ce1/python/vosk/__init__.py#L151

v2: Fixed str=None to str=""
	- vosk_grammar_file could default to "" instead of None, for simplicity.
	- Same for grammar_json.

Signed-off-by: Eric Wheeler <vosk-git@z.ewheeler.org>