Added --verbose flag for "begin" (#85)
Added --verbose=level comand line argument for "begin"

Moved suspend/resume notices to stderr and use --verbose=level

Added Model loading/loaded prints if --verbose is above zero.  This is useful for
large data-sets because it takes a while to load.  Then you can when it is ready
to take dictation.