Add SIGHUP support to reload the user configuration.
The user_config which is typically ~/.config/nerd-dictation/nerd-dictation.py
(or defined via --config) is reloaded when nerd-dictation recieves a SIGHUP.

Syntax error handling:

1. If it is the first time the configuration is being loaded then nerd-dictation
will exit if there is a syntax error in the configuration (maintains original
behavior).

2. If there is a syntax error during reload (SIGHUP), then nerd-dictation will
warn the user, but continue to use the previous configuration so they can fix
the problem without a crash.

Signed-off-by: Eric Wheeler <vosk-git@z.ewheeler.org>