Added auto-grammar correction example that uses Language Tool
This example has minimal configuration and lets Language Tool do the work to
make the dictation look right. (I am not at all affiliated with Language tool,
I just found it useful to work with nerd-dictation.)

Language Tool is a free service (or local server for the paranoid!) for low
request volume and I have added it to my nerd-dictation configuration to
correct and add punctuation where necessary.  I have found that it makes far,
far less manual editing in the result of the spoken text:
  https://languagetool.org/

Signed-off-by: Eric Wheeler <vosk-git@z.ewheeler.org>