Reload --vosk-grammar-file JSON on SIGHUP
This commit adds support for reloading the grammar configuration on SIGHUP.
This is accomplished by moving the recognizer initialization into its own
function so that the signal handler can call it on reload.

A reload of the recognizer is fast because it does not reload the model and
only applies the grammar.

Signed-off-by: Eric Wheeler <vosk-git@z.ewheeler.org>